+++ src/Contracts/WhmcsInterface.php
@@ -4,6 +4,8 @@ namespace Eugenefvdm\Api\Contracts;
 
 interface WhmcsInterface
 {
+    public function addClient(array $data): array;
+
     public function createClientGroup(string $name, string $color = '#ffffff'): void;
 
     public function createCustomClientField(string $name, string $type = 'text'): void;